fix: Shopping cart

+++ Explorer/src/app/feature-modules/marketplace/shopping-cart/shopping-cart.component.ts
@@ -15,6 +15,7 @@ import { Observable } from "rxjs";
 import { ShoppingCart } from "../model/shopping-cart";
 import { PagedResults } from "src/app/shared/model/paged-results.model";
 import { TourToken } from "../model/tour-token.model";
+import { StakeholderService } from "../../stakeholder/stakeholder.service";
 
 @Component({
     selector: "xp-shopping-cart",
@@ -26,9 +27,10 @@ export class ShoppingCartComponent {
     orderItem: OrderItem;
     shoppingCart: ShoppingCart;
     isDisabled: boolean;
-
+    
     constructor(
         private service: MarketplaceService,
+        private stakeholderService: StakeholderService,
         public dialogRef: MatDialog,
         private authService: AuthService,
         public dialog: MatDialogRef<ShoppingCartComponent>,
@@ -93,25 +95,33 @@ export class ShoppingCartComponent {
         var totalPrice=this.shoppingCart.totalPrice
         var storedShoppingCart=this.shoppingCart
         var uslo=false
-        this.service.deleteShoppingKart(this.shoppingCart.id).subscribe({
-            next: () => {
-                this.shoppingCart = {};
-                console.log(storedShoppingCart);
-                this.service.addShoppingCart(this.shoppingCart).subscribe({
-                    next: async (result: ShoppingCart) => {
-                        this.shoppingCart = result;
-                        var newShoppingCart=result;
-                        for (let tour of this.data) {
-                            this.shoppingCart=storedShoppingCart;
-                            const result = await this.service.addToken(tour.id,this.shoppingCart.touristId as number,totalPrice as number);
-                            console.log(result);
-                            alert("You have successfully bought tours!");
-                            this.shoppingCart=newShoppingCart;
-                        }
-                        this.dialogRef.closeAll();
-                    }, 
-                });
-            },
+        this.stakeholderService.getTouristWallet().subscribe(result => {
+            var wallet = result;
+            if(wallet.adventureCoin>=(totalPrice as number)){
+                this.service.deleteShoppingKart(this.shoppingCart.id).subscribe({
+                    next: () => {
+                        this.shoppingCart = {};
+                        console.log(storedShoppingCart);
+                        this.service.addShoppingCart(this.shoppingCart).subscribe({
+                            next: async (result: ShoppingCart) => {
+                                this.shoppingCart = result;
+                                var newShoppingCart=result;
+                                for (let tour of this.data) {
+                                    this.shoppingCart=storedShoppingCart;
+                                    const result = await this.service.addToken(tour.id,this.shoppingCart.touristId as number,totalPrice as number);
+                                    totalPrice=totalPrice as number-tour.price;
+                                    console.log(result);
+                                    alert("You have successfully bought tours!");
+                                    this.shoppingCart=newShoppingCart;
+                                }
+                                this.dialogRef.closeAll();
+                            }, 
+                        });
+                    }
+                })
+            }else{
+                alert("You don't have enough coins.")
+            }
         });
     }
     isEmpty(): void {
